Correct 1-array_iterator.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -12,7 +12,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int index;
+	unsigned int index;
 
 	if ((array == NULL) || (action == NULL))
 	{
